anomaly_detector/requirements.txt
connexion[swagger-ui]
flask-cors
pyyaml
requests
pykafka
